backend/requirements-prod.txt
# 生产环境极简依赖 - 目标：镜像大小 < 200MB
# 只保留绝对必需的核心功能

# Web框架 (核心)
Flask>=2.0.0,<3.0.0
Flask-CORS>=3.0.0,<5.0.0
Werkzeug>=2.0.0,<3.0.0

# 数学计算 (核心)
numpy>=1.19.0,<1.25.0
numpy-financial>=1.0.0,<2.0.0

# 日期处理
python-dateutil>=2.8.0,<3.0.0

# WSGI服务器
gunicorn>=20.0.0,<21.0.0

# Excel导出 (保留，因为是核心功能)
openpyxl>=3.0.0,<4.0.0

# 可选：如果需要pandas DataFrame功能，可以添加
# pandas>=1.3.0,<2.1.0

# 可选：如果需要服务端图表生成，可以添加
# plotly>=5.0.0,<6.0.0  # 相对较轻的图表库

# 移除的重量级依赖：
# - matplotlib (~100MB) -> 图表功能移到前端
# - seaborn (~50MB) -> 统计图表移到前端  
# - scipy (~80MB) -> 如果只做基本计算可以不需要
# - reportlab (~30MB) -> PDF生成功能暂时移除
# - 所有测试和开发工具 (~100MB+)
